[EX1] Added first implementation of GUIActor

+++ ex1/src/main/java/pcd/ass03/model/SimulationMetrics.java
@@ -1,4 +1,5 @@
 package pcd.ass03.model;
 
-public record SimulationMetrics(int totalBoids, double fps,
-                                long tickNumber, double avgNeighbors) {}
+public record SimulationMetrics(int totalBoids, int fps,
+                                long tickNumber, double avgNeighbors) {
+}
